Add more interpretation to the worker healthcheck (#77)
The existing health check for the worker just returns the `inspect ping`
result from celery directly. This means that it's up to the caller to
interpret the results, which is more work than possible on most
container services. This moves the existing health check to
`/health/raw` and changes the default `/health` endpoint to add more
interpretation, returning a 500 error code if there are no healthy
workers locally.