Add a better status & landing page (#95)
- Add a landing page at the root path so that people have a clear
message that the API started correctly and some instructions about what
to do next.
- Add a `/status` page that gives a much more comprehensive overview of
the app's current status. This is intended to be user-facing, unlike the
`/api/v1/health` endpoint on the API and the `/health` endpoint on the
workers, which are meant for programmatic inspection and automated
remediation.
- The status page can return either JSON to be scraped to a metrics
service, or HTML for humans to read.
- The status page will return 200 for healthy services or 500 for
errors. Health is determined essentially by whether we can see any happy
workers. (There might be some unhealthy workers, but as long as there is
at least one healthy one the service won't report errors.)