Encrypt tfstate storage with CMK (#111)
Updates the init script to create an encryption key to use for
encrypting terraform.state blobs.

These changes are backwards-compatible with existing deployments. The
next time any site runs `./backend/init.sh` they will be upgraded to use
the CMK. This is _not_ a required procedure for anyone on GovCloud and
is intended only to accommodate future deployments on Commercial Cloud.

Additional CMKs will need to be provisioned for other resources, this
only affects the terraform.state blob right now.