Support for CMKs in Redis (#116)
Adds support for customer managed keys in Redis.

## Background

Redis is configured to run entirely in memory, however there are certain
circumstances beyond our control where it's possible that data can get
written to disk. This happens for example if the OS is under memory
pressure and will write memory to swap space. Even though this should be
a rare edge case, when it happens, any data written to disk will
normally be encrypted with Azure's managed keys. This is not an issue in
GovCloud, but when running in Commercial cloud we need to guarantee that
any data at rest is protected by a customer-managed key.

Azure's managed Redis service does not offer CMKs. In order to configure
CMKs we need to use Azure's Enterprise Redis Cluster. This is
unfortunately modeled as a completely separate resource than the
Standard Redis SKU, and also isn't fully configurable through the
`azurerm` provider. So this PR does a complete rewrite of the Redis
resources in order to add the option of using the Enterprise SKU with
CMKs.

## Config changes

To use redis enterprise with CMKs, users will need to add the following
lines to their config:

```
redis_sku_family       = "E"
redis_capacity_sku     = 5
```

This will deploy the `Enterprise_E5` SKU. Note that this has less RAM
and more cost than the default `Standard C4` SKU that we deploy. It's
probably only necessary on Commercial Cloud deployments, others should
continue using the default (no config changes required).

## Code changes

This adds support for RedisCluster in celery backend and transport, and
also as a k/v store. The Enterprise SKUs on Azure are clusters, which
have some important differences from the simple redis protocol.