Reject redactions based on quality (#120)
- Add new redaction quality metrics to pipeline
 - Set a configurable max error rate for the pipeline
- Reject redactions when we measure a character error rate higher than
the set maximum

Rejected redactions will be retried per the normal retry policy. When we
reach the maximum number of retries we will fail the redaction task and
set the document status to failed, with the `LowQuality` error as the
descriptor of the failure.

In the event that we fail to generate a high quality redaction, the
client can either re-submit the job to try again or bypass blind review
for this document.

## Future work

- Retries may want to adjust parameters for the redaction request to try
to avoid hallucinations, either with a refined prompt or reduced entropy
or both.
- The specific threshold set in this PR is just a guess about what might
be useful. We want to pick a threshold such that some minor
hallucinations that correct things like grammar and punctuation are
tolerated, while blatant hallucinations are caught and discarded.
Currently we reject documents with a character error rate of 1 in 200
characters, which is fairly generous. We can compute more sophisticated
metrics that consider the actual invalid segments individually and do a
more thorough vetting of the errors.