Rewrite terraform app config to be chunk-aware (#123)
Rewrites the app config in Terraform to account for output window
constraints in the parse and redact/inspect steps and apply a limited
amount of iteration when truncation is detected.

By default the config will attempt up to 5 iterations to complete the
specified task. This limit can be increased by editing the
`{redact,parse}_chunks_max_iterations` variables in the Terraform
variables file.

Since this PR relies on the rewritten experimental chunking support
that's been available in the app for several weeks, deploying this new
Terraform change should not result in any new functions. There will,
however, be a regression (turning chunking support off) if a new API
image is deployed _without_ running the terraform update.