Enable embedding always (#128)
When turned on, embedding (compute + storage) accounts for a small
fraction of one percent of all inference costs. Considering the fixed
infrastructure costs as well, the cost of embedding is a negligible
cost. So we will now always compute embeddings for all documents
submitted for redaction, whether we currently need it or not. This will
give us more information if we need it in the future, with really no
downside even if we never need them.